api_rag/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.1
langchain>=0.1.0
langchain-community>=0.0.21
langchain-core>=0.1.27
langchain-text-splitters>=0.0.1
langchain-ollama>=0.0.1
chromadb>=0.4.22
python-docx>=1.1.0
PyMuPDF>=1.23.26
pandas>=2.2.1
openpyxl>=3.1.2
